split up viz container into subcomponents to make label-flipping easier to implement
