+++ src/visx-visualization/layout/BottomCenter.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import HeatmapXAxis from "../HeatmapXAxis";
+import VisualizationPanel, { VisualizationPanelProps } from "./Panel";
+
+export default function BottomCenterPanel({ id }: VisualizationPanelProps) {
+  return (
+    <VisualizationPanel id={id}>
+      <HeatmapXAxis />
+    </VisualizationPanel>
+  );
+}
