+++ src/visx-visualization/layout/BottomLeft.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import { LeftGraphScale } from "../LeftGraph";
+import VisualizationPanel, { VisualizationPanelProps } from "./Panel";
+
+export default function BottomLeftPanel({ id }: VisualizationPanelProps) {
+  return (
+    <VisualizationPanel id={id}>
+      <LeftGraphScale />
+    </VisualizationPanel>
+  );
+}
