+++ src/visx-visualization/layout/BottomRight.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import VisualizationPanel, { VisualizationPanelProps } from "./Panel";
+
+export default function BottomRightPanel({ id }: VisualizationPanelProps) {
+  return <VisualizationPanel id={id}></VisualizationPanel>;
+}
