+++ src/visx-visualization/layout/MiddleCenter.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import Heatmap from "../Heatmap";
+import VisualizationPanel, { VisualizationPanelProps } from "./Panel";
+
+export default function MiddleCenterPanel({ id }: VisualizationPanelProps) {
+  return (
+    <VisualizationPanel id={id}>
+      <Heatmap />
+    </VisualizationPanel>
+  );
+}
