+++ src/visx-visualization/layout/MiddleLeft.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import LeftGraph from "../LeftGraph";
+import VisualizationPanel, { VisualizationPanelProps } from "./Panel";
+
+export default function MiddleLeftPanel({ id }: VisualizationPanelProps) {
+  return (
+    <VisualizationPanel id={id}>
+      <LeftGraph />
+    </VisualizationPanel>
+  );
+}
