+++ src/visx-visualization/layout/TopRight.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import { TopGraphScale } from "../TopGraph";
+import VisualizationPanel, { VisualizationPanelProps } from "./Panel";
+
+export default function TopRightPanel({ id }: VisualizationPanelProps) {
+  return (
+    <VisualizationPanel id={id}>
+      <TopGraphScale />
+    </VisualizationPanel>
+  );
+}
